Make Javac + Lombok and AspectJ separate modules
First, Javac builds lomboked binaries without any aspects applied. Then,
another module weaves the Spring transaction aspects into the lomboked
binaries and builds the final Spring Boot application.

The application was also modified to catch the deliberately thrown
runtime exception and log it along with all entities. The number of
entities (0 in the transactional, 2 in the non-transactional case) is
the controller's response body. This way, an integration test can easily
verify the behaviour.
