+++ application-lombok/src/main/java/com/example/demo/main/config/MainConfiguration.java
@@ -7,8 +7,5 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 @Configuration
 @EnableTransactionManagement(mode = AdviceMode.ASPECTJ)
-@EnableAutoConfiguration(exclude = {
-        org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration.class
-})
-public class MainConfiguration {
-}
+@EnableAutoConfiguration(exclude = { org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration.class })
+public class MainConfiguration {}
