+++ application-lombok/src/main/java/com/example/demo/main/repository/TestRepository.java
@@ -3,5 +3,4 @@ package com.example.demo.main.repository;
 import com.example.demo.main.entity.TestEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestRepository extends JpaRepository<TestEntity, Long> {
-}
+public interface TestRepository extends JpaRepository<TestEntity, Long> {}
