+++ application-lombok/src/main/resources/application.properties
